langchain-playground/pyproject.toml
[tool.poetry]
name = "langchain-playground"
version = "0.1.0"
description = ""
authors = ["MuhammadAhsaanAbbasi <mahsaanabbasi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langsmith = "^0.1.125"
langserve = "^0.3.0"
langchain-google-genai = "^2.0.0"
langchain-huggingface = "^0.1.0"
langchain-core = "^0.3.5"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.1"
chromadb = "^0.5.7"
tiktoken = "^0.7.0"
sentence-transformers = "^3.1.1"
bs4 = "^0.0.2"
firecrawl-py = "^1.2.4"
langchainhub = "^0.1.21"
wikipedia = "^1.4.0"
tavily-python = "^0.5.0"
langchain-chroma = "^0.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
